feat(front): add canvas drag-and-drop, linking mode, delete dialogs, pending-focus and step-reorder hooks
